Revert "Refactor offer code to keep it DRY in preparation for Enterprise Offers feature."
This reverts commit bc0cb796a853fbfce43cf30b145f6e6492470ad7.
