+++ ecommerce/extensions/api/v2/tests/views/test_baskets.py
@@ -382,10 +382,7 @@ class BasketCalculateViewTests(ProgramTestMixin, TestCase):
     @httpretty.activate
     def test_basket_calculate_program_offer(self):
         """ Verify successful basket calculation with a program offer """
-        offer = ProgramOfferFactory(
-            site=self.site,
-            benefit=PercentageDiscountBenefitWithoutRangeFactory(value=100)
-        )
+        offer = ProgramOfferFactory(benefit=PercentageDiscountBenefitWithoutRangeFactory(value=100))
         program_uuid = offer.condition.program_uuid
         self.mock_program_detail_endpoint(program_uuid, self.site_configuration.discovery_api_url)
         self.mock_enrollment_api(self.user.username)
@@ -462,7 +459,6 @@ class BasketCalculateViewTests(ProgramTestMixin, TestCase):
         self.site_configuration.enable_partial_program = True
         self.site_configuration.save()
         offer = ProgramOfferFactory(
-            site=self.site,
             benefit=PercentageDiscountBenefitWithoutRangeFactory(value=100),
             condition=ProgramCourseRunSeatsConditionFactory()
         )
@@ -494,7 +490,6 @@ class BasketCalculateViewTests(ProgramTestMixin, TestCase):
         self.site_configuration.enable_partial_program = True
         self.site_configuration.save()
         offer = ProgramOfferFactory(
-            site=self.site,
             benefit=PercentageDiscountBenefitWithoutRangeFactory(value=100),
             condition=ProgramCourseRunSeatsConditionFactory()
         )
