+++ ecommerce/programs/templates/programs/programoffer_list.html
@@ -1,6 +1,6 @@
 {% extends 'edx/base.html' %}
 {% load i18n %}
-{% load offer_tags %}
+{% load programs %}
 {% load staticfiles %}
 
 {% block title %}{% trans "Program Offers" %}{% endblock %}
